Internal release for user testing (#13)
* Rooms/create new room (#5)

SSE-15 Implement Create Payment

In this feature, the autogenerated create room got completely 
overhauled and implemented.

It is now possible, to create a new room, which contains of a 
maximal capacity, multiple prices for different accommodations and 
multiple pictures, which each can have a description.

* Bug: Room-Picture and Room Price not stored correctly (#6)

SSE-69 Bug: Pictures and Prices not saved

Due to a programming error, the image and the price for a room was not saved correctly. This was fixed with this PR.

* Create Detail Page for Room (#7)

SSE-67 Detail Page for a Room

In the detail page of a room, the user sees the important data of the room (number, images and prices) and is able to reserve the room.

* Backend: create bookings (#9)

* SSE-20: Implement Booking workflow @ endpoint POST/api/bookings.
- Add DTOs
- Migrate Database
 - Booking hast now a 'main' customer
 - Booking now references RoomPrices instead of Rooms

* SSE-20: Fix failing tests

* Rework the Room Capacity GUI (#8)

SSE-68 Overhaul the Room Capacitiy GUIs to fit the style of the application

As part of room creation, a user must have the possibility to create new capacities to be set for the room. These functionality was provided by JHipster and was now adjusted to fit our application

* SSE-28: cancel booking endpoint added (#10)

* SSE-28: cancel booking endpoint added

* SSE-28: fix tests

* SSE-28: make booking code unique, send email on booking

* SSE-28: no duplicate booking for rooms

* Room/list rooms (#11)

* SSE-17
frontend:
implement frontend
load rooms without price or picture from backend
disable routing for /rooms (click-through prototyp)
use same site for customer and administration

backend:
make /get requests public
route public room requests under /api/public/room

* SSE-17: fix sorting by id

* SSE-17: update tests to reflect api changes

* SSE-17: remove link to reservation and instead always link to room details

* SSE-17: add rooms with prices to fake-data

* SSE-17: load price and picture from backend, add minPrice to possible sort

* SSE-17: make picture api public

* SSE-17: fix test by removing console.log

* SSE-17: fix test by using const instead of let

* SSE-17: revert unwanted changes from testing

* Reservation fe (#12)

* started implementing reservation frontend

Co-authored-by: Dominik Granig <domi.granig@gmail.com>

* SSE28: make booking public

* created reservation model

* SSE28: fix tests

* implemented reservations in fe

* added validation

* SSE28: make delete booking private

* implemented cancel booking

* updated navigation

* updaetd styling

* SSE-20: Price is now influenced by number of days.

* bugfixes and others

Co-authored-by: Dominik Granig <domi.granig@gmail.com>
Co-authored-by: Simon Schwantler <simon.schwantler@powerbot-trading.com>

Co-authored-by: T-DogCodes <44497950+T-DogCodes@users.noreply.github.com>
Co-authored-by: Dominik Granig <domi.granig@gmail.com>
Co-authored-by: 11778256 <58707379+11778256@users.noreply.github.com>
Co-authored-by: DavidJaeger10 <david.jaeger10@gmail.com>